% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_conversion_factor}
\alias{ca_conversion_factor}
\title{Conversion factor dataset}
\format{a \code{tbl_df} with 19,505 observations of the following
4 fields:
\describe{
\item{FoodID}{Sequential number generated by the database identifying the
food records. Joins with \link{ca_food_name}.}
\item{MeasureID}{Sequential number generated by the database for the measure
descriptions. Joins with \link{ca_measure_name}.}
\item{ConversionFactorValue}{The factor by which one would multiply the
nutrient per 100g to obtain nutrient amounts per the measure described (the
weight of that food in the measure described divided by 100).}
\item{ConvFactorDateOfEntry}{	Date the data was entered into the database.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset contains portion size conversion factors. The conversion factors
are food specific multipliers by which the nutrient values for each food may
be multiplied to give the nutrients in described portions.
}
\examples{
ca_conversion_factor

if(require("dplyr")){
head(ca_conversion_factor, 20) \%>\%
left_join(ca_measure_name) \%>\%
left_join(ca_food_name) \%>\%
select(FoodDescription, ConversionFactorValue, MeasureDescription)
}

}
\seealso{
\link{CAnutrients}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
