% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_food_group}
\alias{ca_food_group}
\title{Food group dataset}
\format{a \code{tbl_df} with 23 observations of the following 4 fields:
\describe{
\item{FoodGroupID}{Sequential number generated by the database for the food
groups. Joins with \link{ca_food_name}.}
\item{FoodGroupCode}{Identifier code for the Canadian Nutrient File food
groups. There are 23 different CNF food groups.}
\item{FoodGroupName}{Food group name in English. }
\item{FoodGroupNameF}{Food group name in French.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset is a support or "list" table that is used to link to the
\link{ca_food_name} table. It contains a list of 23 different group headings
(in English and French) based on similar characteristics of the foods.
}
\examples{
ca_food_group

if (require("dplyr")){
head(ca_food_name, 20) \%>\% 
left_join(ca_food_group) \%>\%
select(FoodDescription, FoodGroupName)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
