% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_food_name}
\alias{ca_food_name}
\title{Food name dataset}
\format{a \code{tbl_df} with 5,690 observations of the 10 following fields:
\describe{
\item{FoodID}{Sequential number generated by the database identifying the
food records.  This is the primary key. Joins with \link{ca_nutrient_amount},
\link{ca_conversion_factor}, \link{ca_yield_amount}, and
\link{ca_refuse_amount}.}
\item{FoodCode}{	Identifier code for the foods.  In the 2007b version this
was the FD_C. }
\item{FoodGroupID}{Sequential number generated by the database for the food
groups. Joins with \link{ca_food_group}.}
\item{FoodSourceID}{Sequential number generated by the database for the food
sources. Joins with \link{ca_food_source}.}
\item{FoodDescription}{Complete food name in English.}
\item{FoodDescriptionF}{Complete food name in French.}
\item{FoodDateOfEntry}{Date the food name data was entered into the database}
\item{FoodDateOfPublication}{Date the food name data was entered into the
database.}
\item{CountryCode}{Corresponds to the USDA NDB code.}
\item{ScientificName}{Scientific name of the food.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset stores information about each food in the database. It contains
a description of each food in English and French as well as dates and
comments.
}
\examples{
ca_food_name

if (require("dplyr")){ 
head(ca_food_name, 20) \%>\% 
left_join(ca_food_group) \%>\%
select(FoodDescription, FoodGroupName)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
