% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_food_source}
\alias{ca_food_source}
\title{Food source dataset}
\format{a \code{tbl_df} with 16 observations of the following 4 fields:
\describe{
\item{FoodSourceID}{	Sequential number generated by the database for the food
sources. Joins with \link{ca_food_name}.}
\item{FoodSourceCode}{Identifier code for the food sources. Food sources give
an indication as to Canadian content.}
\item{FoodSourceDescription}{Food Source description in English}
\item{FoodSourceDescriptionF}{Food Source description in French}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset is a support or "list" table that is used to link to the
\link{ca_food_name} table. It contains a list of several food sources (in
English and French) that foods can be grouped on.
}
\examples{
ca_food_source

if (require("dplyr"){
head(ca_food_name, 20) \%>\% 
left_join(food_source) \%>\%
select(FoodDescription, FoodSourceDescription)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
