% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_nutrient_amount}
\alias{ca_nutrient_amount}
\title{Nutrient amount dataset}
\format{a \code{tbl_df} with 524,674 observations of the following 7 fields:
\describe{
\item{FoodID}{Sequential number generated by the database identifying the
food records. Joins with \link{ca_food_name}.}
\item{NutrientID}{Sequential number generated by the database for the
nutrient names. Joins with \link{ca_nutrient_name}.}
\item{NutrientValue}{Mean value in 100g, edible portion. (The number of
decimal places does not reflect the accuracy of the data.)}
\item{StandardError}{Standard error of the mean.}
\item{NumberofObservations}{Number of samples.}
\item{NutrientSourceID}{Identifier code generated by the database for the
nutrient source. Joins with \link{ca_nutrient_source}.}
\item{NutrientDateOfEntry}{Date the data was entered into the database.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This is the main dataset. It uses information (by linking) from the
\link{ca_food_name} table (among others) to identify which nutrients and
amounts are recorded for that food.
}
\examples{
ca_nutrient_amount

if (require("dplyr")){
head(ca_food_name) \%>\%
left_join(ca_nutrient_amount) \%>\%
left_join(ca_nutrient_name) \%>\% 
select(FoodDescription, NutrientName, NutrientValue)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
