% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_nutrient_name}
\alias{ca_nutrient_name}
\title{Nutrient name dataset}
\format{a \code{tbl_df} with 152 observations of the following 8 fields:
\describe{
\item{NutrientID}{Sequential number generated by the database for the
nutrient names. Joins with \link{ca_nutrient_amount}.}
\item{NutrientCode}{Identifier code for nutrient name descriptions.}
\item{NutrientSymbol}{Nutrient symbol or abbreviation of the nutrients. They
may differ from international nomenclature.}
\item{NutrientUnit}{Unit of measure (e.g., mg, g, mcg)}
\item{NutrientName}{Nutrient name in English}
\item{NutrientNameF}{Nutrient name in French}
\item{Tagname}{International Network of Food Data Systems (INFOODS) Tagnames.
A unique Abbreviation for a food component developed by INFOODS to aid in the
interchange of data.}
\item{NutrientDecimals}{	Number of decimal places used in the rounding of the
nutrient value.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset is a support or "list" table that contains the list of nutrients
(in English and French) used in the \link{ca_nutrient_amount} table, with
which it is linked.
}
\examples{
ca_nutrient_name

if (require("dplyr")){
head(ca_food_name) \%>\%
left_join(ca_nutrient_amount) \%>\%
left_join(ca_nutrient_name) \%>\% 
select(FoodDescription, NutrientName, NutrientValue)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
