% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_nutrient_source}
\alias{ca_nutrient_source}
\title{Nutrient source dataset}
\format{a \code{tbl_df} with 20 observations of the following 4 fields:
\describe{
\item{NutrientSourceID}{	Sequential number generated by the database for the
nutrient sources. Joins with \link{ca_nutrient_amount}.}
\item{NutrientSourceCode}{Identifier code for nutrient source descriptions.}
\item{NutrientSourceDescription}{Nutrient source description in English.}
\item{NutrientSourc.DescriptionF}{Nutrient source description in French.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset is a support or "list" table that is used to link to the
\link{ca_nutrient_amount} file. It contains a list of several sources and/or
types of nutrient data (in English and French).
}
\examples{
ca_nutrient_source
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_amount}
\link{ca_yield_name}
}
