% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_refuse_name}
\alias{ca_refuse_name}
\title{Refuse name dataset}
\format{a \code{tbl_df} with 1,047,335 observations of the following 3 fields:
\describe{
\item{RefuseID}{Sequential number generated by the database for the refuse
descriptions. Joins with \link{ca_refuse_amount}.}
\item{RefuseDescription}{Refuse description in English.}
\item{RefuseDescriptionF}{Refuse description in French.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset is a support or "list" table that is used to link to the
\link{ca_refuse_amount} table. It contains a list of refuse types.
}
\examples{
ca_refuse_name

if(require("dplyr")){
  head(ca_refuse_name) \%>\% 
    left_join(ca_refuse_amount) \%>\%
    filter(RefuseDescription != "total refuse") \%>\%
    left_join(ca_food_name) \%>\%
    select(FoodDescription, RefuseDescription)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_yield_amount}
\link{ca_yield_name}
}
