% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{data}
\name{ca_yield_amount}
\alias{ca_yield_amount}
\title{Yield amount dataset}
\format{a \code{tbl_df} with 1,426 observations of the following 4 fields:
\describe{
\item{FoodID}{Sequential number generated by the database identifying the
food records. Joins with \link{ca_food_name}.}
\item{YieldID}{Sequential number generated by the database for the yield
descriptions. Joins with \link{ca_yield_name}.}
\item{YieldAmount}{Yield from refuse and/or cooking losses.}
\item{YieldDateofEntry}{Date the data was entered into the database.}
}}
\source{
The data comes from Canadian Nutrient File 2015.
}
\description{
This dataset contains the yield from refuse and/or cooking losses assigned to
certain foods. These yields are most often used for food inventory purposes.
}
\examples{
ca_yield_amount

if (require("dplyr")){
  head(ca_yield_amount) \%>\% 
    left_join(ca_yield_name) \%>\%
    left_join(ca_food_name) \%>\%
    select(FoodDescription, YieldDescription, YieldAmount)
}
}
\seealso{
\link{CAnutrients}
\link{ca_conversion_factor}
\link{ca_food_group}
\link{ca_food_name}
\link{ca_food_source}
\link{ca_measure_name}
\link{ca_nutrient_amount}
\link{ca_nutrient_name}
\link{ca_nutrient_source}
\link{ca_refuse_amount}
\link{ca_refuse_name}
\link{ca_yield_name}
}
