% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnutrients.R
\docType{package}
\name{CAnutrients}
\alias{CAnutrients}
\alias{CAnutrients-package}
\title{Tidied Canadian Nutrient File 2015 Datasets}
\description{
All eleven datasets from the Canadian Nutrient File (CNF) 2015
edition, tidied for correct data types and removal of extraneous columns.
}
\details{
CAnutrients
A data package for CAnutrients.

Use \code{data(package='CAnutrients')$results[, 3]} to see a list of
available data sets in this data package and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{ca_conversion_factor}, \link{ca_food_group}, \link{ca_food_name},
\link{ca_food_source}, \link{ca_measure_name}, \link{ca_nutrient_amount},
\link{ca_nutrient_name}, \link{ca_nutrient_source}, \link{ca_refuse_amount},
\link{ca_refuse_name}, \link{ca_yield_amount}, \link{ca_yield_name}
}
